requirements.txt
aniso8601==7.0.0
asgiref==3.6.0
Django==4.1.4
django-filter==22.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
djangorestframework==3.14.0
Faker==15.3.4
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
Pillow==9.3.0
print-pp==0.4
promise==2.3
psycopg2==2.9.5
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2022.7
Rx==1.6.3
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tzdata==2022.7
